[#1] refactor: Rename page and component files to 'index'
