+++ src/components/Navigation/index.test.tsx
@@ -3,7 +3,7 @@ import { BrowserRouter } from 'react-router-dom';
 
 import { render, screen } from '@testing-library/react';
 
-import Navigation from './Navigation';
+import Navigation from '.';
 
 it('should render Navigation', () => {
   render(
