+++ src/pages/CountPage/index.test.tsx
@@ -3,7 +3,7 @@ import { RecoilRoot } from 'recoil';
 
 import { render, screen } from '@testing-library/react';
 
-import CountPage from './CountPage';
+import CountPage from '.';
 
 it('should render "Count"', () => {
   render(
