+++ src/pages/HomePage/index.test.tsx
@@ -2,7 +2,7 @@ import React from 'react';
 
 import { render, screen } from '@testing-library/react';
 
-import HomePage from './HomePage';
+import HomePage from '.';
 
 it('should render "Home"', () => {
   render(<HomePage />);
